Return parsing results for incorrect grammars
This refactoring brings the breaking change in the behaviour
- now when grammar or source expression contains errors and
 cannot be parsed request will return 200. This is makes
sense because invalid grammar from parser point of view just
yet another scenario and should not be treated as a failure.
Currently, it doesn't return success indicator or errors
list, but this will be extended soon. This change allows to
save incorrect results as well as corrent.

Still not sure about 200 / 400 on incorrect grammar though.